Add support for running OpenRC within the container.
We need to run multiple processes within the same mount namespace, which
makes using a process manager useful.  To support it correctly, we need to
run /sbin/init within a new mount & PID namespace.  This also means we need
a wrapper script for /sbin/rc-service to execute within the same namespace
(using nsenter).

Signed-off-by: Mark Yen <mark.yen@suse.com>