Install cni-plugins instead of using the nerdctl plugins
We need the "flannel" and "loopback" plugins that are not included
in the nerdctl tarball. The only plugin we extract is "isolation"
because it is not part of the cni-plugins package.

This is needed to run k3s against the system containerd instead of
using the k3s containerd.

This PR also changes the install location from /usr/local/libexec/cni
to /usr/libexec/cni (the default location for containerd).

Signed-off-by: Jan Dubois <jan.dubois@suse.com>