Support iptables-legacy
We upgrade to alpine 3.19 in #83, where iptables has switched to defaulting
to the nftables backend.  This does not work in some WSL scenarios where
the kernel is older than 5.15.

To work around this, install the `iptables-legacy` package, and replace
`/sbin/ip{,6}tables{,-save,-restore}` with a shell script that detects the
kernel version and arranges for the correct backend to be called.

Signed-off-by: Mark Yen <mark.yen@suse.com>